+++ cloudfront_geo_restriction/outputs.tf
@@ -1,0 +1,7 @@
+output "restriction_type" {
+  value = var.enabled ? "blacklist" : "none"
+}
+
+output "locations" {
+  value = var.enabled ? local.locations : []
+}