+++ cloudfront_geo_restriction/variables.tf
@@ -1,0 +1,5 @@
+variable "enabled" {
+  type = bool
+  default = true
+  description = "Enables geo-restrictions. Disabled module still outputs valid values that apply no restrictions."
+}