+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [1.0.29] - 2022-07-07
 
-- [Domain] Add geo-restrictions.
+- [CloudFront Geo-Restriction] Add helper module for projects and other modules that define CloudFront distributions.
+- [Domain] Add geo-restriction.
+- [Static Site] Add geo-restriction.
 
 ## [1.0.28] - 2022-06-24